refactor player and ray code
