+++ C/raycast/raycast/raycast.vcxproj
@@ -144,6 +144,8 @@
     <ClCompile Include="Graphics.c" />
     <ClCompile Include="Main.c" />
     <ClCompile Include="Map.c" />
+    <ClCompile Include="Player.c" />
+    <ClCompile Include="Ray.c" />
     <ClCompile Include="Textures.c" />
     <ClCompile Include="upng.c" />
   </ItemGroup>
@@ -151,6 +153,8 @@
     <ClInclude Include="Definitions.h" />
     <ClInclude Include="Graphics.h" />
     <ClInclude Include="Map.h" />
+    <ClInclude Include="Player.h" />
+    <ClInclude Include="Ray.h" />
     <ClInclude Include="Textures.h" />
     <ClInclude Include="TextureVerts.h" />
     <ClInclude Include="upng.h" />
